splitout datastore tests by type
